% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRMTLd.R
\name{aRMTLd}
\alias{aRMTLd}
\title{Title Comparing the difference of adjusted restricted mean time lost (aRMTLd) under competing risks}
\usage{
aRMTLd(time, status, group, covariates, tau)
}
\arguments{
\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 0 = right censored, 1 = event of interest, 2 = competing event.}

\item{group}{The group indicator for comparison. The elements of this vector take either 1 or 0.
Normally, 0 = control group, 1 = treatment group.}

\item{covariates}{A matrix of covariates.}

\item{tau}{The value to specify the truncation time point for the RMTL calculation.}
}
\value{
The value of aRMTLd, the stand error and 95% confidence interval of aRMTLd。
}
\description{
Title Comparing the difference of adjusted restricted mean time lost (aRMTLd) under competing risks
}
\examples{
library(aRMTL)
data(simdata)
time=simdata$time;status=simdata$status;group=simdata$group
covariates <- cbind(simdata$z1,simdata$z2,simdata$z3,simdata$z4)
aRMTLd(time,status,group,covariates,tau = 365)
}
