% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRMTLd_curve.R
\name{aRMTLd_curve}
\alias{aRMTLd_curve}
\title{Title The difference in adjusted restricted mean time lost (aRMTLd) curves under competing risks}
\usage{
aRMTLd_curve(B, tL, tU, time, status, group, covariates)
}
\arguments{
\item{B}{The number of random variables generated of standard normal distribution.}

\item{tL}{The beginning of aRMTLd curve time interval.}

\item{tU}{The end of aRMTLd curve time interval.}

\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 0 = right censored, 1 = event of interest, 2 = competing event.}

\item{group}{The group indicator for comparison. The elements of this vector take either 1 or 0.
Normally, 0 = control group, 1 = treatment group.}

\item{covariates}{A matrix of covariates.}
}
\value{
A six-column data box, where tt is listed as the limit time point,
diff is listed as the aRMTLd, ci_lower and ci_upper are the upper and lower bounds of the confidence interval,
and cb_lower and cb_upper are the lower bounds of the confidence band.
}
\description{
Title The difference in adjusted restricted mean time lost (aRMTLd) curves under competing risks
}
\examples{
library(aRMTL)
data(simdata)
time=simdata$time;status=simdata$status;group=simdata$group
covariates <- cbind(simdata$z1,simdata$z2,simdata$z3,simdata$z4)
set.seed(2024)
B=10
tL <- ceiling(max(min(time[group==0&status==1]), min(time[group==1&status==1])))
tU <- 365
result <- aRMTLd_curve(B,tL,tU,time,status,group,covariates)
plot(result$tt,result$diff,type = "l",ylim = c(-105,35),xlim = c(0,365),xlab = "time(days)" , ylab = "aRMTLd",
col = "brown3",main = "Group1 - Group0: aRMTLd curve plot",lwd=3)
lines(result$tt,result$ci_lower,type = "l",col = "brown3",lwd=3,lty=2)
lines(result$tt,result$ci_upper,type = "l",col = "brown3",lwd=3,lty=2)
polygon(c(result$tt,rev(result$tt)),c(result$cb_lower,rev(result$cb_upper)),
col=rgb(1,0,0,0.15),border=FALSE)
legend("topleft", legend = "aRMTLd", col=c("brown3"),  lwd=3, cex=1, bty ="n", inset = c(0, 0))
abline(h=0,col = "gray30",lwd=3,lty=3)
}
